c2py-agent/requirements.txt
pycparser>=2.21
openai>=1.0.0
pytest>=7.0.0
typed-ast>=1.5.4
black>=23.0.0
isort>=5.12.0
pylint>=2.17.0
pyyaml>=6.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0 